requirements.txt
alembic==0.9.4
click==6.7
factory-boy==2.9.2
Faker==0.7.18
Flask==0.12.2
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
py==1.4.34
pytest==3.1.2
python-dateutil==2.6.1
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
